requirements.txt
spotipy>=2.23.0
ytmusicapi>=1.7.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
